test: reuse existing rdbms restore logic
