+++ zeebe/qa/integration-tests/src/test/java/io/camunda/zeebe/it/cluster/backup/RdbmsRangeRestoreIT.java
@@ -7,206 +7,51 @@
  */
 package io.camunda.zeebe.it.cluster.backup;
 
-import static io.camunda.cluster.PhysicalTenantIds.DEFAULT_PHYSICAL_TENANT_ID;
-import static io.camunda.configuration.beanoverrides.BrokerBasedPropertiesOverride.RDBMS_EXPORTER_NAME;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.awaitility.Awaitility.await;
 
-import io.camunda.application.commons.rdbms.RdbmsDataSources;
-import io.camunda.client.CamundaClient;
 import io.camunda.configuration.Camunda;
 import io.camunda.configuration.Filesystem;
 import io.camunda.configuration.PrimaryStorageBackup.BackupStoreType;
-import io.camunda.configuration.SecondaryStorage.SecondaryStorageType;
-import io.camunda.db.rdbms.sql.ExporterPositionMapper;
-import io.camunda.db.rdbms.write.RdbmsMapperBundle;
-import io.camunda.management.backups.BackupInfo;
-import io.camunda.management.backups.PartitionBackupInfo;
-import io.camunda.management.backups.StateCode;
-import io.camunda.zeebe.model.bpmn.Bpmn;
-import io.camunda.zeebe.qa.util.actuator.BackupActuator;
-import io.camunda.zeebe.qa.util.actuator.ExportersActuator;
 import io.camunda.zeebe.qa.util.cluster.TestRestoreApp;
-import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
-import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
-import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.util.FileUtil;
 import java.nio.file.Path;
-import java.sql.SQLException;
-import java.time.Duration;
-import java.time.Instant;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
-import org.awaitility.Awaitility;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.util.unit.DataSize;
 
 /**
- * Integration test that verifies restoring from a time range of backups when RDBMS is configured as
- * the secondary storage. This exercises the RDBMS-aware restore path in {@code RestoreManager}
- * where {@code ExporterPositionMapper} is used by {@code BackupRangeResolver} to determine safe
- * restore points per partition.
- *
- * <p>Uses a filesystem backup store (no containers needed) and H2 in-memory database for RDBMS.
+ * Restores from a time range of backups via a separate, standalone restore application ({@code
+ * TestRestoreApp}). See {@link RdbmsRangeRestoreTestBase} for the shared fixture and test cases,
+ * and {@link InProcessRdbmsRangeRestoreIT} for the in-process counterpart.
  */
-@ZeebeIntegration
-final class RdbmsRangeRestoreIT implements ClockSupport {
-
-  private static final Logger LOG = LoggerFactory.getLogger(RdbmsRangeRestoreIT.class);
-  // H2 in-memory DB URL shared across broker and restore app JVM lifecycles.
-  // DB_CLOSE_DELAY=-1 keeps the DB alive after last connection closes.
-  // This is necessary so that the broker can restart without losing the exported data
-  // MODE=PostgreSQL for SQL compatibility with the RDBMS module.
-  private static final String H2_URL =
-      "jdbc:h2:mem:rdbms-restore-" + UUID.randomUUID() + ";DB_CLOSE_DELAY=-1;MODE=PostgreSQL";
+final class RdbmsRangeRestoreIT extends RdbmsRangeRestoreTestBase {
 
   private static @TempDir Path backupDir;
+
   private Path workingDirectory;
-
-  @TestZeebe
-  private final TestStandaloneBroker broker =
-      new TestStandaloneBroker()
-          .withSecondaryStorageType(SecondaryStorageType.rdbms)
-          .withUnifiedConfig(this::configureBroker);
-
-  private BackupActuator backupActuator;
-  private ExporterPositionMapper exporterPositionMapper;
 
   @BeforeEach
   void setUp() {
-    backupActuator = BackupActuator.of(broker);
     workingDirectory = broker.getWorkingDirectory();
-    final Map<String, RdbmsMapperBundle> bundles = broker.bean("rdbmsMapperBundles");
-    final var defaultBundle = bundles.get(DEFAULT_PHYSICAL_TENANT_ID);
-    if (defaultBundle == null) {
-      throw new IllegalStateException(
-          "Missing default physical tenant '%s' in rdbmsMapperBundles; known tenants: %s"
-              .formatted(DEFAULT_PHYSICAL_TENANT_ID, bundles.keySet()));
-    }
-    exporterPositionMapper = defaultBundle.exporterPositionMapper();
-    pinClock(broker);
   }
 
-  @AfterEach
-  void tearDown() throws SQLException {
-    final var dataSources = broker.bean(RdbmsDataSources.class);
-    if (dataSources != null) {
-      final var ds = dataSources.dataSourceFor(DEFAULT_PHYSICAL_TENANT_ID);
-      try (final var connection = ds.getConnection()) {
-        connection.createStatement().execute("DROP ALL OBJECTS");
-      }
-    }
+  @Override
+  protected Path backupDir() {
+    return backupDir;
   }
 
-  @Test
-  void shouldNotRestoreIfMissingBackupAfterRestorePosition() throws Exception {
-    // given - deploy a process and create instances, then take continuous backups.
-    final long processKey;
-
-    final Interval interval;
-    try (final var client = broker.newClientBuilder().build()) {
-      processKey = deployTestProcess(client);
-
-      // Create some process instances to have data to verify after restore
-      interval = createProcessInstancesAndTakeBackups(client, processKey);
-
-      takeAndAwaitBackup();
-      progressClock(broker, 2000);
-      client.newCreateInstanceCommand().processDefinitionKey(processKey).send().join();
-      client.newCreateInstanceCommand().processDefinitionKey(processKey).send().join();
-    }
-
-    // when - stop broker, delete data, restore from time range with RDBMS
+  @Override
+  void restoreFromTimeRange(final Interval interval) throws Exception {
     broker.stop();
-    FileUtil.deleteFolder(workingDirectory);
-    FileUtil.ensureDirectoryExists(workingDirectory);
-
-    // then
-    try (final var restoreApp = testRestoreApp(interval)) {
-      assertThatThrownBy(restoreApp::start)
-          .isInstanceOf(IllegalStateException.class)
-          .hasMessageContaining("No usable range found");
-    }
-  }
-
-  @Test
-  void shouldRestoreFromATimeRange() throws Exception {
-    // given - some process instances with backups between them
-    final long processKey;
-
-    final Interval interval;
-    try (final var client = broker.newClientBuilder().build()) {
-      processKey = deployTestProcess(client);
-      interval = createProcessInstancesAndTakeBackups(client, processKey);
-    }
-    takeAndAwaitBackup();
-
-    final var exporterActuator = ExportersActuator.of(broker);
-    exporterActuator.disableExporter(RDBMS_EXPORTER_NAME);
-
-    Awaitility.await("Until backup is greater or equal to exported position")
-        .pollDelay(Duration.ofSeconds(2))
-        .atMost(Duration.ofSeconds(300))
-        .untilAsserted(
-            () -> {
-              final var backup = takeAndAwaitBackup();
-              final var position = exporterPositionMapper.findOne(1).lastExportedPosition();
-              final var details = backup.getDetails().getFirst();
-              assertThat(details)
-                  .returns(1, PartitionBackupInfo::getPartitionId)
-                  .returns(StateCode.COMPLETED, PartitionBackupInfo::getState);
-              assertThat(details.getCheckpointPosition()).isGreaterThanOrEqualTo(position);
-            });
-
-    // when - stop broker, delete data, restore from time range with RDBMS
-    broker.stop();
-
     FileUtil.deleteFolder(workingDirectory);
     FileUtil.ensureDirectoryExists(workingDirectory);
     try (final var restore = testRestoreApp(interval)) {
       restore.start();
     }
-
     broker.start();
-
-    completeJobs(4);
   }
 
-  @Test
-  void shouldRestoreWithoutArguments() throws Exception {
-    // given - deploy a process and create instances, then take continuous backups.
-    final long processKey;
-
-    try (final var client = broker.newClientBuilder().build()) {
-      processKey = deployTestProcess(client);
-      createProcessInstancesAndTakeBackups(client, processKey);
-    }
-    takeAndAwaitBackup();
-
-    final var exporterActuator = ExportersActuator.of(broker);
-    exporterActuator.disableExporter(RDBMS_EXPORTER_NAME);
-
-    Awaitility.await("Until backup is greater or equal to exported position")
-        .pollDelay(Duration.ofSeconds(2))
-        .atMost(Duration.ofSeconds(300))
-        .untilAsserted(
-            () -> {
-              final var backup = takeAndAwaitBackup();
-              final var position = exporterPositionMapper.findOne(1).lastExportedPosition();
-              final var details = backup.getDetails().getFirst();
-              assertThat(details)
-                  .returns(1, PartitionBackupInfo::getPartitionId)
-                  .returns(StateCode.COMPLETED, PartitionBackupInfo::getState);
-              assertThat(details.getCheckpointPosition()).isGreaterThanOrEqualTo(position);
-            });
-    // when - stop broker, delete data, restore from time range with RDBMS
+  @Override
+  void restoreWithoutArguments() throws Exception {
     broker.stop();
     FileUtil.deleteFolder(workingDirectory);
     FileUtil.ensureDirectoryExists(workingDirectory);
@@ -214,127 +59,28 @@ final class RdbmsRangeRestoreIT implements ClockSupport {
       restore.start();
     }
     broker.start();
-
-    completeJobs(4);
   }
 
-  private void completeJobs(final int expectedJobCount) {
-    try (final var client = broker.newClientBuilder().build()) {
-      int tries = 10;
-      int remaining = expectedJobCount;
-      while (remaining > 0 && tries-- > 0) {
-        final var jobActivationResult =
-            client
-                .newActivateJobsCommand()
-                .jobType("task")
-                .maxJobsToActivate(remaining)
-                .send()
-                .join();
-        remaining -= jobActivationResult.getJobs().size();
-        for (final var job : jobActivationResult.getJobs()) {
-          client.newCompleteCommand(job.getKey()).send().join();
-        }
-      }
+  @Override
+  void assertRestoreFailsForMissingBackup(final Interval interval) throws Exception {
+    broker.stop();
+    FileUtil.deleteFolder(workingDirectory);
+    FileUtil.ensureDirectoryExists(workingDirectory);
+    try (final var restoreApp = testRestoreApp(interval)) {
+      assertThatThrownBy(restoreApp::start)
+          .isInstanceOf(IllegalStateException.class)
+          .hasMessageContaining("No usable range found");
     }
-  }
-
-  private static long deployTestProcess(final CamundaClient client) {
-    return client
-        .newDeployResourceCommand()
-        .addProcessModel(
-            Bpmn.createExecutableProcess("process")
-                .startEvent()
-                .serviceTask("task", t -> t.zeebeJobType("task"))
-                .endEvent()
-                .done(),
-            "process.bpmn")
-        .send()
-        .join()
-        .getProcesses()
-        .getFirst()
-        .getProcessDefinitionKey();
-  }
-
-  private Interval createProcessInstancesAndTakeBackups(
-      final CamundaClient client, final long processKey) {
-
-    // Create some process instances to have data to verify after restore
-    client.newCreateInstanceCommand().processDefinitionKey(processKey).send().join();
-    client.newCreateInstanceCommand().processDefinitionKey(processKey).send().join();
-
-    takeAndAwaitBackup();
-    // Record "from" before the first backup, with a small sleep to create clear separation
-
-    final var from = currentTime(broker);
-    LOG.info("Time range from = {}", from);
-
-    client.newCreateInstanceCommand().processDefinitionKey(processKey).send().join();
-    // Take first backup
-    takeAndAwaitBackup();
-
-    // Create another instance and take second backup
-    client.newCreateInstanceCommand().processDefinitionKey(processKey).send().join();
-    takeAndAwaitBackup();
-
-    // Record "to" after second backup
-    final Instant to = currentTime(broker);
-    LOG.info("Time range to = {}", to);
-    return new Interval(from, to);
-  }
-
-  private BackupInfo takeAndAwaitBackup() {
-    progressClock(broker, 2000);
-    final var res = backupActuator.take();
-    final long backupId = Optional.ofNullable(res.getBackupId()).orElseThrow();
-
-    await("backup is completed")
-        .timeout(Duration.ofSeconds(30))
-        .ignoreExceptions()
-        .untilAsserted(
-            () ->
-                assertThat(backupActuator.status(backupId).getState())
-                    .isEqualTo(StateCode.COMPLETED));
-
-    return backupActuator.status(backupId);
-  }
-
-  private void configureBroker(final Camunda cfg) {
-    cfg.getSystem().setClockControlled(true);
-    configureRdbms(cfg);
-    // Filesystem backup store
-    final var fsConfig = new Filesystem();
-    fsConfig.setBasePath(backupDir.toAbsolutePath().toString());
-    cfg.getData().getPrimaryStorage().getBackup().setFilesystem(fsConfig);
-    cfg.getData().getPrimaryStorage().getBackup().setStore(BackupStoreType.FILESYSTEM);
-
-    // Enable continuous backups
-    cfg.getData().getPrimaryStorage().getBackup().setContinuous(true);
-
-    // Use smaller log segments for faster checkpoint creation.
-    // Max message size must be <= log segment size (enforced by RaftPartitionFactory).
-    cfg.getData().getPrimaryStorage().getLogStream().setLogSegmentSize(DataSize.ofMegabytes(1));
-    cfg.getCluster().getNetwork().setMaxMessageSize(DataSize.ofKilobytes(500));
-
-    // Disable auto-configure camunda exporter to avoid conflicts
-    cfg.getData().getSecondaryStorage().setAutoconfigureCamundaExporter(false);
   }
 
   private void configureRestoreApp(final Camunda cfg) {
     configureRdbms(cfg);
     // Filesystem backup store (same as broker)
     final var fsConfig = new Filesystem();
-    fsConfig.setBasePath(backupDir.toAbsolutePath().toString());
+    fsConfig.setBasePath(backupDir().toAbsolutePath().toString());
     cfg.getData().getPrimaryStorage().getBackup().setFilesystem(fsConfig);
     cfg.getData().getPrimaryStorage().getBackup().setStore(BackupStoreType.FILESYSTEM);
     cfg.getData().getPrimaryStorage().getBackup().setContinuous(true);
-  }
-
-  private static void configureRdbms(final Camunda cfg) {
-    cfg.getData().getSecondaryStorage().setType(SecondaryStorageType.rdbms);
-    final var rdbms = cfg.getData().getSecondaryStorage().getRdbms();
-    rdbms.setUrl(H2_URL);
-    rdbms.setUsername("sa");
-    rdbms.setPassword("");
   }
 
   @SuppressWarnings("resource")
@@ -353,6 +99,4 @@ final class RdbmsRangeRestoreIT implements ClockSupport {
         .withUnifiedConfig(this::configureRestoreApp)
         .withWorkingDirectory(workingDirectory);
   }
-
-  private record Interval(Instant start, Instant end) {}
 }
